chore: consolidate 3d code under single module
