+++ src/3d/world/clutter/definitions.js
