+++ src/3d/world/generation/smoke.js
