+++ src/views/primary/WorldMap.vue
@@ -101,14 +101,14 @@ import { ShaderPass } from 'three/examples/jsm/postprocessing/ShaderPass.js';
 import { FXAAShader } from 'three/examples/jsm/shaders/FXAAShader.js';
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
 import api from '@/functions/api';
-import { BIOME_THRESHOLDS } from '@/terrain/biomes';
-import WorldGrid from '@/world/WorldGrid';
-import PlayerMarker from '@/renderer/PlayerMarker';
-import ClutterManager from '@/world/ClutterManager';
-import createRealisticWaterMaterial from '@/renderer/materials/RealisticWaterMaterial';
+import { BIOME_THRESHOLDS } from '@/3d/terrain/biomes';
+import WorldGrid from '@/3d/world/WorldGrid';
+import PlayerMarker from '@/3d/renderer/PlayerMarker';
+import ClutterManager from '@/3d/world/ClutterManager';
+import createRealisticWaterMaterial from '@/3d/renderer/materials/RealisticWaterMaterial';
 import { useSettingsStore } from '@/stores/settingsStore';
-import { updateRadialFadeUniforms } from '@/renderer/radialFade';
-import ChunkManager from '@/renderer/ChunkManager';
+import { updateRadialFadeUniforms } from '@/3d/renderer/radialFade';
+import ChunkManager from '@/3d/renderer/ChunkManager';
 import WorldDebugPanel from '@/components/world/WorldDebugPanel.vue';
 import { profiler, createWebGLTimer } from '@/utils/profiler';
 import { useWorldStore } from '@/stores/worldStore';
